Move models from advanced KB settings to new AI models page
