+++ apps/dashboard/src/app/knowledge-box/knowledge-box-home/knowledge-box-home.component.html
@@ -131,7 +131,7 @@
             }
             @if (config['semantic_model']; as semanticModel) {
               <div>
-                <div class="title-xxs">{{ 'kb.settings.semantic.label' | translate }}</div>
+                <div class="title-xxs">{{ 'kb.ai-models.semantic.label' | translate }}</div>
                 <div class="body-s">
                   {{ semanticModel }}
                 </div>
