+++ libs/common/src/lib/ai-models/index.ts
@@ -1,0 +1,1 @@
+export * from './ai-models.component';
